snresnet us are float64

+++ dis_models/resblocks.py
@@ -25,10 +25,10 @@ class Block(chainer.Chain):
             self.c2 = SNConvolution2D(hidden_channels, out_channels, ksize=ksize, pad=pad, initialW=initializer)
             if self.learnable_sc:
                 self.c_sc = SNConvolution2D(in_channels, out_channels, ksize=1, pad=0, initialW=initializer_sc)
-                self.c_sc.u = np.ones((1, out_channels)).astype(np.float32)
-                
-            self.c1.u = np.ones((1, 128)).astype(np.float32)
-            self.c2.u = np.ones((1, 128)).astype(np.float32)
+                self.c_sc.u = np.ones((1, out_channels)).astype(np.float64)
+
+            self.c1.u = np.ones((1, 128)).astype(np.float64)
+            self.c2.u = np.ones((1, 128)).astype(np.float64)
             
 
     def residual(self, x):
@@ -66,9 +66,9 @@ class OptimizedBlock(chainer.Chain):
             self.c2 = SNConvolution2D(out_channels, out_channels, ksize=ksize, pad=pad, initialW=initializer)
             self.c_sc = SNConvolution2D(in_channels, out_channels, ksize=1, pad=0, initialW=initializer_sc)
 
-            self.c1.u = np.ones((1, 128)).astype(np.float32)
-            self.c2.u = np.ones((1, 128)).astype(np.float32)
-            self.c_sc.u = np.ones((1, 128)).astype(np.float32)
+            self.c1.u = np.ones((1, 128)).astype(np.float64)
+            self.c2.u = np.ones((1, 128)).astype(np.float64)
+            self.c_sc.u = np.ones((1, 128)).astype(np.float64)
 
 
     def residual(self, x):
